Reorganize files for next step
