+++ packages/php-wasm/xdebug-bridge/public/xdebug-bridge.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-import './index.js';
+import './cli.js';
