+++ packages/php-wasm/xdebug-bridge/src/lib/dbgp-session.ts
